requirements_minimal.txt
# Core C2 Framework Dependencies
flask==3.0.0
flask-cors==4.0.0
cryptography==41.0.7
requests==2.31.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Configuration
pyyaml==6.0.1

# Utilities
python-dotenv==1.0.0
colorama==0.4.6
click==8.1.7
